Refactor: Move SignupViewModel and add PlannerActivityViewModel
This commit refactors the `SignupViewModel` by moving it to a new package and renames it. It also introduces `PlannerActivityViewModel` to manage planner activities data. Additionally, it updates the `PlannerActivityLocalDataSource` interface and its implementation to use suspend functions, and modifies the service locator to add coroutine dispatchers. Also updates `InitialLoadingFragment` and `HomeFragment` and `SignupFragment` to use new view model name.
Here's a summary of the specific changes:

-   Moves `SignupViewModel` from `presenter.signup` to `presenter.viewmodel`.
-   Adds a new `PlannerActivityViewModel` class in `presenter.viewmodel`.
-   Updates `PlannerActivityLocalDataSource` interface and implementation to use `suspend` functions.
-   Adds `ioDispatcher` and `mainDispatcher` to the service locator.
-   Updates `InitialLoadingFragment` and `HomeFragment` and `SignupFragment` to use new view model name.
- Adds the insert function in the `PlannerActivityDao`