Refactor: Implement UpdatePlannerActivityDialog and add delete button
This commit introduces the `UpdatePlannerActivityDialog` to allow users to modify planner activity details.
It also adds a delete button inside the dialog.
In the plannerActivity item, add click listener to open the dialog, and also implement update isCompleted option.
Additionally it was refactored some code to improve the app quality.

Changes include:

-   Add `UpdatePlannerActivityDialog` layout
-   Add `btnDelete` on the `UpdatePlannerActivityDialog` layout
-   Implement `onClickPlannerActivity` inside `PlannerActivityAdapter` to open `UpdatePlannerActivityDialog`
-   Implement `onChangeIsCompleted` inside `PlannerActivityAdapter` to change activity completed status
-   Refactor `PlannerActivityViewModel`
- Refactor `ivStatus` to `ivIsCompleted` and some ids inside the item layout.